Refactor: split landing and pricing pages into modular components
Extract landing page sections (Hero, Features, Testimonials, Partners
Preview, CTA) into src/components/landing/ and pricing sections (Hero,
Tiers, FAQ) into src/components/pricing/. Reduces page.tsx from 600
to 60 lines and pricing/page.tsx from 500 to 162 lines. No file
exceeds 230 lines.
